+++ client/src/styles/theme.ts
@@ -21,7 +21,7 @@ const font = {
     },
     h2: {
       weight: weight.extraLight,
-      size: '1.5rem',
+      size: '2rem',
       lineHeight: '56px',
     },
     h3: {