+++ client/CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Changed mobile layout of place page ([#432](https://github.com/project-lux/lux-frontend/issues/432)).
 - Changed mobile layout of person page ([#435](https://github.com/project-lux/lux-frontend/issues/435)).
 - Changed mobile layout of event page ([#436](https://github.com/project-lux/lux-frontend/issues/436)).
+- Changed mobile layout of concept page ([#437](https://github.com/project-lux/lux-frontend/issues/437)).
 
 ### Removed