+++ client/CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Changed mobile layout of place page ([#432](https://github.com/project-lux/lux-frontend/issues/432)).
+- Changed mobile layout of person page ([#435](https://github.com/project-lux/lux-frontend/issues/435)).
 
 ### Removed